Create CODEOWNERS
bsieber: I’m game for whichever path, I’ve not done a round robin or had it assigned (automatically at least) is this something we can setup in the settings?

bstack: yeah it's a setting on on a github team and then you assign a team as a reviewer in CODEOWNERS

bsieber: I started to tweak the auto-assign settings for this in the it-integrations team; didn’t set a codeowners file just yet though (removed managers from rotation)
…but maybe there’s a better way to organize this, be it a new team; I’m all ears if you’ve got any suggestions.

bstack: either way seems to make sense to me. the tc team over-rotated on this a lot when this was released with really specific segmentation in codeowners etc
bstack: a lot of which was my fault for being too excited about this tooling :)
bstack: so something simple that serves our purposes makes sense to me. Basic desires from my end are:
1) not having to pick someone every time
2) evenly distributing load and familiarity
3) relatively quick feedback even when one person is busy or pto, etc

bstack:
I forgot it was something you need to set up at a team level rather than just a repo level so I don't want to have too much input into your teams internal workings