Add /identities endpoint, query by alt ids
The private /identities endpoint allows querying by the alternate IDs:
 * primary_email
 * amo_user_id
 * basket_token
 * fxa_id
 * fxa_primary_email
 * sfdc_id

Because some alternate IDs are not unique, the return is a list of
IdentityResponses. If none are found, an empty list is returned rather
than a 404.

For completeness, email_id is also allowed, and may be useful when
querying by multiple IDs. If multiple IDs are passed, then all must
match to return the contact.

If no IDs are passed, a 400 Bad Request is returned.

It is currently implemented against the built-in sample contacts.
In the future, it can be transitioned to query the database.