Add /ctms endpoint to query by alternate IDs
The public /ctms endpoint allows querying by primary and alternate IDs:
 * email_id
 * primary_email
 * amo_user_id
 * basket_token
 * fxa_id
 * fxa_primary_email
 * sfdc_id

This works in a similar way to /identies, except the return is a list of
ContactSchema objects. If none are found, the list is empty, and, since
some IDs are non-unique, multiple entries may be returned as well.i

If multiple IDs are passed, then only contacts that match all IDs are
returned. If no IDs are passed, a 400 (Bad Request) is returned.

It is currently implemented against the built-in sample contacts. In the
future, it can be transitioned to query the database.