Run unit tests in a database session transaction
Add a series of fixtures that create a test database, initialize the
schema, and then run each test in a transaction that is rolled back at
the end of the test. The database session is also published to the app,
so that the same session is queried when using the test client.