Convert docker-compose to use environment files
Environment files like .env are considered after the environment
settings in docker-compose.yml and the shell. The
docker/config/local_dev.env file has settings for the local development
environment, such as the database URL. The .env file has
developer-specific settings, is created when a make command is used,
and is ignored by git.

To demonstrate possibilities, I've added a few override for Makefile
behaviour:

* MK_WITH_SERVICE_PORTS - Add --service-ports to calls to
  "docker-compose run", to make it more likely that postgres host
  ports will stay open.
* MK_KEEP_DOCKER_UP - Avoid calling "docker-compose down" after
  "make test"

The real use cases are in the future - adding a dev value for the
per-deployment CTMS_SECRET_KEY for crypto operations, adding a switch
for development vs live servers, etc.