Add OAuth2 client credentials grant
Add a /token endpoint that implements the OAuth2 client credentials
grant process. A caller passes a client_id and a client_secret, either
in the form body or a header, and gets an OAuth2 token that expires.
The token is implemented as a JSON Web Token, signed by the server (HMAC
with SHA-256). When the token expires, the client calls /token again.

All API endpoints now require a bearer token. The Swagger API docs
include a mechanism for authenticating in the browser.

This adds new application settings:

* CTMS_SECRET_KEY - A long string used in JWT signing. There is a
  default for development, and it should be unique for each deployment.
* CTMS_TOKEN_EXPIRATION - How long, in seconds, that access tokens are
  valid. Default is 60 minutes.