Add client_credentials.py
This script is used to generate API client credentials. It takes a short
name to identify the client and a contact email, and generates a
client_secret. It prints out the details of the credentials, which can
be sent to the requester.

It can also be used to update an existing API client, including updating
the email address and disabling the credentials.