Run as user app in docker
Create a user app (default uid=10001, gid=10001) that owns the code and
files and runs the application.

This may cause issues for Linux users, if a command in the container
creates a file with permissions different than the developer's. The .env
settings CTMS_UID and CTMS_GID can be used to change from the defaults.

This is not necessary on Docker Desktop for MacOS or Windows, which uses
a different (slower) approach for binding mounts in Docker.