Allow testing outside of docker
The docker container is in UTC so the timestamps it generates are in
that timezone. The `datetime.utcnow()...timestamp()` in tests is kinda
funny. The note in
https://docs.python.org/3/library/datetime.html#datetime.datetime.timestamp
explains a bit more. Basically when I tried running pytest on my machine
(which is in PST currently), the result of that call was 7 hours in the
past instead of correctly in UTC.