Move MoFo data to own group, add contact ID
Move the Mozilla Foundation data to a new subgroup "mofo", to allow
omitting the data when not set, and to be more flexible about future
additions and changes.

The group has these items:
* mofo_email_id: MoFo's identifier for the email, was email.mofo_id
* mofo_contact_id: MoFo's identifier for a contact, which may have
  several emails
* mofo_relevant: This contact is relevant to MoFo, was
  email.mofo_relevant

mofo_email_id and mofo_contact_id are available as alternate IDs.

The database columns email.mofo_id and email.mofo_relevant will be
removed after this code is in production.

Because we don't have production data yet, the data migration is
omitted.