Add PATCH /ctms/{email_id}
PATCH performs a partial update of the items passed in the JSON body. It
can change any value except for the email_id.

Tests are in a new file, because pylint was complaining about the long
file length of test_api.py, and I agreed that it made sense to break it
up.