Add init_sentry() function, call globally
Add init_sentry() to initialize Sentry. This handles the ValidationError
raised when a required setting is omitted. It could eventually live in a
shared module used by background tasks and cronjobs as well.

It is called at the module level, because FastAPI uses threads to
integrate async and sync code, and this ensures that each thread has
initialized Sentry.

It should be skipped for tests, so avoid installing if "pytest" is in
the command.