Add Sentry as middleware, rather than wrapping app
Starlette, the ASGI framework used by FastAPI, includes an outer
middleware ServerErrorMiddleware that may interfere with Sentry's
exception handlers. Instead of wrapping the FastAPI app as sentry_app,
add SentryAsgiMiddleware as a middleware to the FastAPI app, to try to
handle the error before ServerErrorMiddleware.

The wrapped name 'sentry_app' is built into Docker, so a `make build` is
required.