Delete all but latest duplicate customers
Turn the failing migration 47c59393f5d8 into an empty migration, and
then add two migrations.

The first looks for stripe_customers with duplicate email_ids, ranks
them by Stripe creation time, and deletes all but the most recent one.

The second is the same as the failed migration, but avoids nulled
columns. This may not be necessary.