Trace Stripe API calls when Customer email matches
If the Customer email has the tracing string "+trace-me-mozilla-", then
log the Stripe customer object.

This does not detect if the linked customer / contact should be traced
via a matching email, so it will not trace subscription changes or
invoices, or if a matching Contact has a different but tracing email.