Fix relationships to Stripe models
The relationship between FirefoxAccount and StripeCustomer is not a
foreign key constraint, so the remote / foreign designation swaps
depending on which side you are starting from. They can non
back-populate and are view-only (can't change the relationship via the
relationship).

The relationship from an Email to a StripeCustomer goes through
FirefoxAccount. The remote / foreign designation also swaps, as well as
which join is primary and which is secondary. However, it is able to
back populate. Weird.