Add healthcheck file and script for sync service
Support two optional environment variables:

* BACKGROUND_HEALTHCHECK_PATH - path to a healthcheck file, like
  "/tmp/sync_timestamp"
* BACKGROUND_HEALTHCHECK_AGE_S - maximum age in seconds, like 60

If the path is configured, the acoustic_sync.py script will write the
current time in Python's ISO 8601 format.

The new healthcheck_sync.py script will read this file and check how
long ago the timestamp was written. If it is recent enough, it exits
with code 0. If anything goes wrong, it logs an exception and exits with
code 1.

This is designed to work as a "liveness" check for Kubernetes.