Remove assertions about the text of SQL statements
This is an implementation detail in the context of the ingest_invoice_data
function. We should assert that the `for_update` kwarg does what we
want, and that we don't encounter any integrity errors when trying to do
updates, but asserting certain SQL fragments are in the statements feels
too brittle.