Always run container with single Uvicorn process
Previously, we'd run our application container with uvicorn in
development, but gunicorn with multiple processes in production. This
made the app a little more complex when it came to metrics collection,
and we have k8s to manage scaling. So, with this commit, we remove
gunicorn and now always run the application with a single process.