Always run container with single Uvicorn process (#503)
Previously, we'd run our application container with uvicorn in
development, but gunicorn with multiple workers in production. This
made the app a little more complex when it came to metrics collection,
and we have k8s to manage scaling. So, with this commit, we remove
gunicorn and now always run the application with a single process.

With removing Gunicorn, we also had to do some tweaking to our logging 
and application settings.