Start to use factories to generate test data (#554)
We use static data as test data throughout our tests. Some observations I've made about this data and the tests that use them:
 - it makes test setup not obvious
 - they promote the use of "magic numbers"
 - they're used in complex fixtures
 - The division fixtures, for example "minimal" and "maximal" contact, seems arbitrary

In this commit, we would introduce the use of test factories with `Factory Boy` and `pytest-factoryboy`. 

We've done minimal factory setup in this commit to demonstrate a proof of concept. This involved adding a `scoped_session` for database tests as well as making two example factories and using them in two example tests.

As another minor change, we also added an `__init__.py` to the unit test directory so imports worked correctly