Move sample contact data to pytest fixtures (#628)
* Migrates test data from the custom ContactVendor class in `sample_data.py`
into pytest fixtures. There are two fixtures for each sample -- the
contact with a `_data` suffix that just returns the data in a `ContactSchema`,
and without the `_data` suffix where the contact is persisted to the database

* Uses these new fixtures in tests  to replace the `SAMPLE_CONTACTS` data

* In a few tests, we were getting all of the `sample_contacts`, then only using one
of them in the test. Now, we just get the fixture we want directly.

* use `request.getfixturevalue` to get contact fixtures in parametrized tests

Instead of using the `sample_contacts` fixture then using the
parametrized value to select the correct fixture for the test case, use
request.getfixturevalue. This makes it so we don't have the unexpected
side effect of creating 3 contacts when we only need one

* Remove dbsession fixture in some tests where it's unnecessary