Allow for more testing flexibility for local dev (#718)
Move all logic required for running the test suite into the bin/test.sh
script so that we have more flexibility for how we run the tests.

We can still run `make test`, which will run all of the tests and
produce a coverage report like before. But now, we can _also_ run
`bin/test.sh` to enable the coverage report manually (disabled by
default) and supply pytest args for more flexibility in how we run the
test suite for local development.