Fix special boolean text values on the main table (#804)
These values that we sync to the main contact table of Acoustic

- "double_opt_in",
- "has_opted_out_of_email",
- "fxa_account_deleted",
- "amo_email_opt_in",

are booleans in our database.

We were writing them as `1`/ `0` to Acoustic until
https:/github.com/mozilla-it/ctms-api/pull/762, where we started
recording values as `Yes` / `No`, which is how boolean values are
recorded in Acoustic elsewhere. The queries built on top of the
Acoustic contact database expected `1` / `0`, though.

Until we can migrate the columns and queries that use these colums to proper
boolean types in Acoustic, we're going to temporarily handle these special cases
and backfill `Yes`/ `No` values to their `1` / `0` notation.