Serialize timestamps consistently
The underlying Rust library that handles timestamps apparently uses `Z`
to denote UTC time, rather that a 00:00 offset. Though the Z is "more
correct (I think), Python's isoformat method uses +00:00. Since that's
what most of our timestamps serialized to, we use a custom serializer to
ensure datetimes are serialized consistently.

I do think we shold eventally use the Z offset, but that would be a
breaking change of the API, so deserves to be part of a major release.